proto -- (minor) fixed bug with the Number_reprs
When the value was 0, the Number_reprs (Number_repr, Int_repr, Real_repr)
would return "Number", "Int", or "Real" instead of the actual number. This
is because Python defaults 0 to mean false, and Python also defaults None
to mean false, so by implicitly checking for None, we were also
implicitly checking for 0. Now the check for None has been made
explicit, so we are no longer return "Int" (or whatever) instead of "0".