Moved from mixed name/index slots to all index.
That is, the slot system before had two separate kinds of slots: named
and indexd. Now, they are combined. All slots are indexed, some slots
are also named.

This makes it much easier to iterate over an object. It also makes it
easier to determine the correct order of evaluation of things.