Implemented lexical analyzer (lexer, token).
Starting to switch over to Dekorte's Object-Oriented C style, because
it helps me keep the complexity of the API low. Generally, it pushes
you to have a very simple API, and it makes optional arguments pretty
easy to add later. Functions with many arguments tend to be refactored
into being their own separate "class". This is what happened with
FugaLexer: originally there was only FugaToken and FugaTokenCursor,
but the API was a little too complicated. Now, there is FugaLexer and
FugaToken, and they are all GC'd, and they are easy to use.