Parsing is complete.
A flaw in tools/make has come to my attention. I will need to debug
it eventually. Sometimes it does not rebuild everything when a
dependency is changed. This happened when I add FUGA_TOKEN_EQUALS
to token.h -- it didn't rebuild parser.c. I added #include "token.h"
into parser.c for good measure, but I haven't seen if it "fixes" it
or not.