A large rewrite (not from scratch).
This rewrite's main effect was rearranging how data is stored in each
Fuga object. Now, we refer to all objects as void*. Previously, they
were all Fuga*. Using void* makes sense because now primitive data is
accessed directly at the pointer address, so you can do
Fuga_clone_(proto, size) as a replacement for FugaGC_alloc(gc, size),
and then use the result as your primitive. Object metadata is now
stored before the pointer, as a prefix or header.

This means we have unified garbage collection and the object system.
The big advantage is we have made the system easier to use.

Up next: updating the parser / lexer to the new system, getting
rid of the FugaGC module (since that functionality is now covered
by the Fuga module).